add module with generic cache
