+++ src/client/cache.rs
@@ -1,54 +1,14 @@
-use std::{
-  num::NonZeroUsize,
-  sync::{Arc, RwLock},
-  time::{Duration, Instant},
-};
+use std::sync::Arc;
 
-use lru::LruCache;
+use crate::cache::ResultCache;
+
 use mime::Mime;
 use reqwest::header::HeaderMap;
 use url::Url;
 
 use crate::util::{Error, Result};
 
-struct Timed<T> {
-  value: T,
-  created: Instant,
-}
-
-pub struct ResponseCache {
-  map: RwLock<LruCache<Url, Timed<Response>>>,
-  timeout: Duration,
-}
-
-impl ResponseCache {
-  pub fn new(max_entries: usize, timeout: Duration) -> Self {
-    let max_entries = max_entries.try_into().unwrap_or(NonZeroUsize::MIN);
-    Self {
-      map: RwLock::new(LruCache::new(max_entries)),
-      timeout,
-    }
-  }
-
-  pub fn get_cached(&self, url: &Url) -> Option<Response> {
-    let mut map = self.map.write().ok()?;
-    let entry = map.get(url)?;
-    if entry.created.elapsed() > self.timeout {
-      map.pop(url);
-      return None;
-    }
-    Some(entry.value.clone())
-  }
-
-  pub fn insert(&self, url: Url, response: Response) -> Option<()> {
-    let timed = Timed {
-      value: response,
-      created: Instant::now(),
-    };
-    self.map.write().ok()?.push(url, timed);
-    Some(())
-  }
-}
+pub type ResponseCache = ResultCache<Url, Response>;
 
 #[derive(Clone)]
 pub struct Response {
