+++ src/main.rs
@@ -1,3 +1,4 @@
+mod cache;
 mod cli;
 mod client;
 mod config;
